App changed from ts to js
