+++ README.md
@@ -1,79 +1,5 @@
-# Vue Movie Quiz Project
+# Vue 3 + Vite
 
-==========================
+This template should help get you started developing with Vue 3 in Vite. The template uses Vue 3 `<script setup>` SFCs, check out the [script setup docs](https://v3.vuejs.org/api/sfc-script-setup.html#sfc-script-setup) to learn more.
 
-## Overview
-
-This project is a Vue.js application that provides a movie quiz game. The game is built using Vue 3, TypeScript, and Vite, and features a user-friendly interface with PrimeVue components.
-
-## Features
-
--  Movie quiz game with multiple questions
--  User-friendly interface with PrimeVue components
--  State management and data fetching with Vue Query
--  Routing with Vue Router
--  Styling with Tailwind CSS
--  Auto-importing of Vue and VueUse elements with unplugin-auto-import
--  Auto-importing of components with unplugin-vue-components
-
-## Installation
-
-To get started with the project, follow these steps:
-
-1. Clone the repository: `git clone https://github.com/your-username/vue-movie-quiz.git`
-2. Install the dependencies: `npm install` or `yarn install`
-3. Start the development server: `npm run dev` or `yarn dev`
-
-## Project Structure
-
-The project is organized into the following folders:
-
--  `src`: The source code folder, containing the Vue application code.
--  `public`: The public folder, containing static assets and the index.html file.
--  `utils`: The utilities folder, containing helper functions and configurations.
-
-## Components
-
-The project uses the following components:
-
--  `App.vue`: The main application component.
--  `Home.vue`: The home page component.
--  `About.vue`: The about page component.
--  `TriviaTitle.vue`: The trivia title component.
--  `TriviaDrawer.vue`: The trivia drawer component.
-
-## Routing
-
-The project uses Vue Router for client-side routing. The routes are defined in the `src/main.ts` file.
-
-## State Management
-
-The project uses Vue Query for state management and data fetching.
-
-## Styling
-
-The project uses Tailwind CSS for styling.
-
-## Auto-Importing
-
-The project uses unplugin-auto-import to auto-import Vue and VueUse elements, and unplugin-vue-components to auto-import components.
-
-## Development
-
-To start the development server, run `npm run dev` or `yarn dev`. This will start the Vite development server, and you can access the application at `http://localhost:3000`.
-
-## Building
-
-To build the application for production, run `npm run build` or `yarn build`. This will create a production-ready build of the application in the `dist` folder.
-
-## Deployment
-
-To deploy the application, you can use a hosting platform such as Vercel or Netlify. You can also deploy the application to a custom server or CDN.
-
-## Contributing
-
-Contributions are welcome! If you'd like to contribute to the project, please fork the repository and submit a pull request.
-
-## License
-
-The project is licensed under the MIT License.
+Learn more about IDE Support for Vue in the [Vue Docs Scaling up Guide](https://vuejs.org/guide/scaling-up/tooling.html#ide-support).
