--- components.d.ts
@@ -1,37 +1,0 @@
-/* eslint-disable */
-// @ts-nocheck
-// Generated by unplugin-vue-components
-// Read more: https://github.com/vuejs/core/pull/3399
-export {}
-
-/* prettier-ignore */
-declare module 'vue' {
-  export interface GlobalComponents {
-    AutoComplete: typeof import('primevue/autocomplete')['default']
-    Avatar: typeof import('primevue/avatar')['default']
-    Button: typeof import('primevue/button')['default']
-    Drawer: typeof import('primevue/drawer')['default']
-    FloatLabel: typeof import('primevue/floatlabel')['default']
-    Home: typeof import('./src/components/Home.vue')['default']
-    InputGroup: typeof import('primevue/inputgroup')['default']
-    InputText: typeof import('primevue/inputtext')['default']
-    Menu: typeof import('primevue/menu')['default']
-    Menubar: typeof import('primevue/menubar')['default']
-    ProgressBar: typeof import('primevue/progressbar')['default']
-    ProgressSpinner: typeof import('primevue/progressspinner')['default']
-    QuestionNavigation: typeof import('./src/components/QuestionNavigation.vue')['default']
-    Quiz: typeof import('./src/components/Quiz.vue')['default']
-    RouterLink: typeof import('vue-router')['RouterLink']
-    RouterView: typeof import('vue-router')['RouterView']
-    Select: typeof import('primevue/select')['default']
-    SpeedDial: typeof import('primevue/speeddial')['default']
-    Toast: typeof import('primevue/toast')['default']
-    ToggleButton: typeof import('primevue/togglebutton')['default']
-    TriviaDrawer: typeof import('./src/components/TriviaDrawer.vue')['default']
-    TriviaTitle: typeof import('./src/components/TriviaTitle.vue')['default']
-  }
-  export interface ComponentCustomProperties {
-    Ripple: typeof import('primevue/ripple')['default']
-    StyleClass: typeof import('primevue/styleclass')['default']
-  }
-}
