+++ package.json
@@ -1,11 +1,11 @@
 {
-   "name": "v-movie-quiz",
+   "name": "v-movie-quiz-js",
    "private": true,
    "version": "0.0.0",
    "type": "module",
    "scripts": {
       "dev": "vite",
-      "build": "vue-tsc -b && vite build",
+      "build": "vite build",
       "preview": "vite preview"
    },
    "dependencies": {
@@ -19,20 +19,15 @@
       "lodash": "^4.17.21",
       "primeicons": "^7.0.0",
       "primevue": "^4.0.5",
-      "vue": "^3.4.38",
-      "vue-router": "^4.4.3"
+      "vue": "^3.5.10",
+      "vue-router": "^4.4.5"
    },
    "devDependencies": {
-      "@primevue/auto-import-resolver": "^4.0.5",
-      "@vitejs/plugin-vue": "^5.1.2",
+      "@vitejs/plugin-vue": "^5.1.4",
       "autoprefixer": "^10.4.20",
-      "postcss": "^8.4.41",
-      "tailwindcss": "^3.4.10",
-      "typescript": "^5.5.3",
-      "unplugin-auto-import": "^0.18.2",
-      "unplugin-vue-components": "^0.27.4",
-      "vite": "^5.4.1",
-      "vite-plugin-vue-devtools": "^7.3.9",
-      "vue-tsc": "^2.0.29"
+      "postcss": "^8.4.47",
+      "tailwindcss": "^3.4.13",
+      "vite": "^5.4.8",
+      "vite-plugin-vue-devtools": "^7.3.9"
    }
 }
