+++ tailwind.config.js
@@ -1,6 +1,6 @@
 /** @type {import('tailwindcss').Config} */
 export default {
-   content: ['./index.html', './src/**/*.{vue,js,ts,jsx,tsx}'],
+   content: ['./index.html', './src/**/*.{js,ts,jsx,tsx,vue}'],
    theme: {
       extend: {},
    },
