+++ vite.config.js
@@ -1,0 +1,12 @@
+import { defineConfig } from 'vite'
+import vue from '@vitejs/plugin-vue'
+
+// https://vitejs.dev/config/
+export default defineConfig({
+   plugins: [vue()],
+   resolve: {
+      alias: {
+         '@': '/src',
+      },
+   },
+})
